mysqltest - use my_popen()/my_fgets() wrappers for popen()/fgets()
On Windows, CRT version of popen/fgets() has a bug, such that
fgets()sometimes loses end of line when reading from pipe
(details in MDEV-9409)

The workaround is to use binary mode for the FILE* returned by popen(),
and to strip '\r' from the end of line manually.

The workaround was used previously for "exec", but now wrapper functions
are used also for "perl" snippets.